fix: 총 공부시간 오버플로우 수정 - 총 공부시간을 int에서 long으로 수정

+++ src/main/java/develop/grassserver/common/utils/duration/DurationUtils.java
@@ -36,7 +36,7 @@ public class DurationUtils {
         return String.format("%d:%02d:%02d", hours, minutes, seconds);
     }
 
-    public static int formatHourDuration(Duration duration) {
-        return (int) duration.getSeconds() / 3600;
+    public static long formatHourDuration(Duration duration) {
+        return (long) duration.getSeconds() / 3600;
     }
 }
