+++ src/main/java/develop/grassserver/grass/presentation/dto/YearlyGrassResponse.java
@@ -4,6 +4,6 @@ public record YearlyGrassResponse(
         Long id,
         int month,
         int day,
-        int studyHour
+        long studyHour
 ) {
 }
